pyproject.toml
[project]
name = "mlflow-jupyter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "gcsfs>=2025.10.0",
    "google-cloud-storage>=3.4.1",
    "google-generativeai>=0.8.5",
    "imbalanced-learn>=0.14.0",
    "ipykernel>=7.1.0",
    "ipython>=9.7.0",
    "jedi>=0.19",
    "jupyterlab>=4.2,<4.3",
    "jupyterlab-lsp>=5.1",
    "lightgbm>=4.6.0",
    "litellm[gemini]>=1.80.0",
    "matplotlib>=3.10.7",
    "mlflow[extras]>=3.6.0",
    "notebook-intelligence[jupyterlab]>=2.6.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "python-lsp-server[all]>=1.11",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
]
